create utils directory in testing
